# openapi.json
{"openapi":"3.1.0","info":{"title":"Club Management API","description":"Quản lý câu lạc bộ","version":"1.0"},"servers":[{"url":"http://localhost:8080/club","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"User API","description":"Quản lý các user"}],"paths":{"/api/auth/signup":{"post":{"tags":["authentication-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["authentication-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticationResponse"}}}}}}},"/api/users":{"get":{"tags":["User API"],"summary":"Get tất cả các user trong hệ thống","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserResponse"}}}}}}}},"components":{"schemas":{"UserCreationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8},"fullName":{"type":"string"}}},"ApiResponseUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserResponse"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"status":{"type":"string"},"role":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ApiResponseAuthenticationResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthenticationResponse"}}},"AuthenticationResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserResponse"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"authenticated":{"type":"boolean"}}},"ApiResponseListUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}